perf: Refactor code with HOCs
