+++ JestEnvironment.js
@@ -2,6 +2,8 @@ const NodeEnvironment = require('jest-environment-node');
 
 class HTMLTextAreaElement {}
 
+class CustomEvent {}
+
 class ResizeObserver {
   observe () {}
   unobserve () {}
@@ -20,11 +22,13 @@ class JestEnvironment extends NodeEnvironment {
     await super.setup();
 
     this.global.HTMLTextAreaElement = HTMLTextAreaElement;
+    this.global.CustomEvent = CustomEvent;
     this.global.ResizeObserver = ResizeObserver;
     this.global.getComputedStyle = getComputedStyle;
 
     this.global.window = {
       HTMLTextAreaElement,
+      CustomEvent,
       ResizeObserver,
       getComputedStyle
     };
@@ -33,6 +37,7 @@ class JestEnvironment extends NodeEnvironment {
   async teardown () {
     await super.teardown();
     this.global.HTMLTextAreaElement = undefined;
+    this.global.CustomEvent = undefined;
     this.global.ResizeObserver = undefined;
     this.global.getComputedStyle = undefined;
     this.global.window = undefined;
